feat: move safeJSON to util
